wip another go at v13
this thing does not work as intended with svelte because the svelte
code is executed on the client. It needs to make a request to the
server to get the image. But the loadbalancer portforwarfing fucks
this up. E.g. we need to hardcode the port.

To fix: either rewrite to only use the node backend, or that the
node backend can also serve the image upon request using the
cached image.
I think that is how i should rewrite. The image request should
always get the image, not the filepath to the image as the file-
path is in the container and inaccessible in client side.