WIP ex3.01,
Error with getting the correct baseurl. I think the proper way is
to not hardcode it. I've deleted therefrore (but not yet built and
deployed containers) the call to config which obtain localhost as
env var. also deleted the config endpoint from backend.
I guess this should work

what is not working is the caching of the picsum image. No idea yet
why.